move generating federation config to graphql pkg
